feat(fe-215): gate rewards and achievements routes behind env feature flags
- Add FEATURE_FLAGS module driven by NEXT_PUBLIC_FEATURE_REWARDS/ACHIEVEMENTS
- Extract AchievementsView client component; achievements/page.tsx becomes server component with flag guard
- Add /rewards route with flag guard using RewardsClaimPanel
- Add ComingSoon placeholder for disabled routes
- Document flags in .env.example
- Add vitest spec for flag logic

Closes #753
