+++ frontend/.env.example
@@ -1,1 +1,5 @@
 NEXT_PUBLIC_API_URL=https://dewordle.onrender.com/api/v1
+
+# Feature flags — set to "true" to enable incomplete surfaces
+NEXT_PUBLIC_FEATURE_REWARDS=false
+NEXT_PUBLIC_FEATURE_ACHIEVEMENTS=false
